centralize the sha256 coprocessor implementation
