+++ examples/sha256_nivc.rs
@@ -1,34 +1,22 @@
-use std::marker::PhantomData;
-use std::sync::Arc;
-use std::time::Instant;
-
-use lurk::circuit::circuit_frame::MultiFrame;
-use lurk::circuit::gadgets::data::GlobalAllocations;
-use lurk::circuit::gadgets::pointer::{AllocatedContPtr, AllocatedPtr};
-use lurk::coprocessor::{CoCircuit, Coprocessor};
-use lurk::eval::{empty_sym_env, lang::Lang};
-use lurk::field::LurkField;
-use lurk::proof::{supernova::SuperNovaProver, Prover};
-use lurk::ptr::Ptr;
-
-use lurk::public_parameters::instance::{Instance, Kind};
-use lurk::public_parameters::{public_params_default_dir, supernova_public_params};
-use lurk::state::user_sym;
-use lurk::store::Store;
-use lurk::tag::{ExprTag, Tag};
-use lurk::Num;
-use lurk_macros::Coproc;
-
-use bellpepper::gadgets::multipack::pack_bits;
-use bellpepper::gadgets::sha256::sha256;
-use bellpepper_core::boolean::Boolean;
-use bellpepper_core::{ConstraintSystem, SynthesisError};
-
 use pasta_curves::pallas::Scalar as Fr;
-use serde::{Deserialize, Serialize};
-use sha2::{Digest, Sha256};
+use std::{sync::Arc, time::Instant};
 use tracing_subscriber::{fmt, prelude::*, EnvFilter, Registry};
 use tracing_texray::TeXRayLayer;
+
+use lurk::{
+    circuit::circuit_frame::MultiFrame,
+    coprocessor::sha256::{Sha256Coproc, Sha256Coprocessor},
+    eval::{empty_sym_env, lang::Lang},
+    field::LurkField,
+    proof::{supernova::SuperNovaProver, Prover},
+    ptr::Ptr,
+    public_parameters::{
+        instance::{Instance, Kind},
+        public_params_default_dir, supernova_public_params,
+    },
+    state::user_sym,
+    store::Store,
+};
 
 const REDUCTION_COUNT: usize = 10;
 
@@ -61,115 +49,6 @@ fn sha256_nivc<F: LurkField>(store: &mut Store<F>, n: usize, input: Vec<usize>) 
     );
 
     store.read(&program).unwrap()
-}
-
-#[derive(Clone, Debug, Serialize, Deserialize)]
-pub(crate) struct Sha256Coprocessor<F: LurkField> {
-    n: usize,
-    pub(crate) _p: PhantomData<F>,
-}
-
-impl<F: LurkField> CoCircuit<F> for Sha256Coprocessor<F> {
-    fn arity(&self) -> usize {
-        self.n
-    }
-
-    fn synthesize<CS: ConstraintSystem<F>>(
-        &self,
-        cs: &mut CS,
-        _g: &GlobalAllocations<F>,
-        _store: &Store<F>,
-        input_exprs: &[AllocatedPtr<F>],
-        input_env: &AllocatedPtr<F>,
-        input_cont: &AllocatedContPtr<F>,
-    ) -> Result<(AllocatedPtr<F>, AllocatedPtr<F>, AllocatedContPtr<F>), SynthesisError> {
-        let zero = Boolean::constant(false);
-
-        let mut bits = vec![];
-
-        // println!("{:?}", input_exprs);
-
-        for input_ptr in input_exprs {
-            let tag_bits = input_ptr
-                .tag()
-                .to_bits_le_strict(&mut cs.namespace(|| "preimage_tag_bits"))?;
-            let hash_bits = input_ptr
-                .hash()
-                .to_bits_le_strict(&mut cs.namespace(|| "preimage_hash_bits"))?;
-
-            bits.extend(tag_bits);
-            bits.push(zero.clone()); // need 256 bits (or some multiple of 8).
-            bits.extend(hash_bits);
-            bits.push(zero.clone()); // need 256 bits (or some multiple of 8).
-        }
-
-        bits.reverse();
-
-        let mut digest_bits = sha256(cs.namespace(|| "digest_bits"), &bits)?;
-
-        digest_bits.reverse();
-
-        // Fine to lose the last <1 bit of precision.
-        let digest_scalar = pack_bits(cs.namespace(|| "digest_scalar"), &digest_bits)?;
-        let output_expr = AllocatedPtr::alloc_tag(
-            &mut cs.namespace(|| "output_expr"),
-            ExprTag::Num.to_field(),
-            digest_scalar,
-        )?;
-        Ok((output_expr, input_env.clone(), input_cont.clone()))
-    }
-}
-
-impl<F: LurkField> Coprocessor<F> for Sha256Coprocessor<F> {
-    fn eval_arity(&self) -> usize {
-        self.n
-    }
-
-    fn simple_evaluate(&self, s: &Store<F>, args: &[Ptr<F>]) -> Ptr<F> {
-        let mut hasher = Sha256::new();
-
-        let mut input = vec![0u8; 64 * self.n];
-
-        for (i, input_ptr) in args.iter().enumerate() {
-            let input_zptr = s.hash_expr(input_ptr).unwrap();
-            let tag_zptr: F = input_zptr.tag().to_field();
-            let hash_zptr = input_zptr.value();
-            input[(64 * i)..(64 * i + 32)].copy_from_slice(&tag_zptr.to_bytes());
-            input[(64 * i + 32)..(64 * (i + 1))].copy_from_slice(&hash_zptr.to_bytes());
-        }
-
-        input.reverse();
-
-        hasher.update(input);
-        let mut bytes = hasher.finalize();
-        bytes.reverse();
-        let l = bytes.len();
-        // Discard the two most significant bits.
-        bytes[l - 1] &= 0b00111111;
-
-        let scalar = F::from_bytes(&bytes).unwrap();
-        let result = Num::from_scalar(scalar);
-
-        s.intern_num(result)
-    }
-
-    fn has_circuit(&self) -> bool {
-        true
-    }
-}
-
-impl<F: LurkField> Sha256Coprocessor<F> {
-    pub(crate) fn new(n: usize) -> Self {
-        Self {
-            n,
-            _p: Default::default(),
-        }
-    }
-}
-
-#[derive(Clone, Debug, Coproc, Serialize, Deserialize)]
-enum Sha256Coproc<F: LurkField> {
-    SC(Sha256Coprocessor<F>),
 }
 
 /// Run the example in this file with
